backend/requirements.txt
fastapi
uvicorn
pydantic
transformers
torch
sqlalchemy
psycopg2-binary
python-dotenv
requests